[rename] register/fields.ts -> register/const.ts
