+++ apps/fe/src/app/register/const.ts
